Fix_userId data type

+++ src/main/java/KEPCO/SSD/device/repository/DeviceRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface DeviceRepository extends JpaRepository<Device, String> {
-    List<Device> findByUserId(int userId); // int로 변경
-    Optional<Device> findByUserIdAndSerialNumber(int userId, String serialNumber); // int로 변경
-    void deleteByUserIdAndSerialNumber(int userId, String serialNumber); // int로 변경
+    List<Device> findByUserId(long userId);
+    Optional<Device> findByUserIdAndSerialNumber(long userId, String serialNumber);
+    void deleteByUserIdAndSerialNumber(long userId, String serialNumber);
 }
